+++ integrations/pyroscope/src/error.rs
@@ -1,0 +1,12 @@
+// Copyright 2019 TiKV Project Authors. Licensed under Apache-2.0.
+
+#[derive(Debug, thiserror::Error)]
+pub enum Error {
+    #[error("{0}")]
+    ReqwestError(#[from] reqwest::Error),
+
+    #[error("{0}")]
+    PProfError(#[from] pprof::Error),
+}
+
+pub type Result<T> = std::result::Result<T, Error>;
